refactor(ingredient): extract post-parse passes into refine.rs
Pure move, no behavior change. Lifts postprocess_ingredient and its six
ordered passes, plus the modifier helpers (append_modifier, clean_modifier,
strip_wrapping_parens) and extraction helpers (extract_alternative,
extract_secondary_amounts) and the strip_wrapping_parens test, out of
pipeline.rs into refine.rs. pipeline.rs now holds only orchestration, the
nom grammar, and the nom->Ingredient builder glue (949 -> ~250 lines).
