+++ ingredient-parser/src/parser/mod.rs
@@ -7,6 +7,7 @@ pub(crate) mod measurement;
 pub(crate) mod normalize;
 pub(crate) mod pipeline;
 pub(crate) mod recognize;
+pub(crate) mod refine;
 
 pub(crate) use helpers::parse_amount_string;
 pub(crate) use helpers::{
